requirements.txt
# Core dependencies
psycopg2-binary>=2.9.9  # PostgreSQL database adapter
python-dotenv>=1.0.0    # Environment variable management
pyyaml>=6.0.2          # YAML processing

# Testing and Development
pytest>=7.4.3          # Testing framework
pytest-asyncio>=0.21.1 # Async test support
pytest-cov>=4.1.0      # Coverage reporting
pytest-mock>=3.12.0    # Mocking support
psutil>=6.1.1          # Process utilities

# Type checking and Linting
pyright>=1.1.335       # Static type checking
mypy>=1.7.0           # Type checking
pylint>=3.0.2         # Code analysis

# Documentation
sphinx>=7.2.6         # Documentation generation
sphinx-rtd-theme>=1.3.0  # Documentation theme
